refactor(repositories): split message_repository into focused modules
Separate types/trait, persist mapping (error-column envelope), index
metadata, and SQLite query orchestration while keeping public exports stable.
