--- src-tauri/src/repositories/message_repository.rs
@@ -1,1232 +1,0 @@
-use super::error::DbError;
-use crate::models::chat::{Message, MessageSource};
-use crate::utils::pagination::Page;
-use async_trait::async_trait;
-use sea_orm::{
-    sea_query::Expr, ColumnTrait, ConnectionTrait, DatabaseBackend, DatabaseConnection,
-    EntityTrait, PaginatorTrait, QueryFilter, QueryResult, QuerySelect, Set, Statement,
-};
-
-use crate::entity::prelude::{Message as MessageEntity, MessageIndexMeta};
-use crate::entity::{message, message_index_meta, session};
-use crate::utils::json::{from_json_option, from_json_or_default, to_json_option};
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct MessagePaginationCursor {
-    pub created_at: i64,
-    pub row_id: i64,
-}
-
-#[derive(Debug, Clone)]
-pub struct MessageSlicePage {
-    pub items: Vec<Message>,
-    pub has_more_before: bool,
-    pub oldest_cursor: Option<MessagePaginationCursor>,
-}
-
-#[derive(Debug)]
-struct MessageRowWithCursor {
-    model: message::Model,
-    cursor: MessagePaginationCursor,
-}
-
-/// Message repository trait for abstraction and testability
-#[async_trait]
-pub trait MessageRepository: Send + Sync {
-    /// Retrieve a paginated list of messages for a specific session
-    async fn get_page(
-        &self,
-        session_id: &str,
-        page: u64,
-        page_size: u64,
-    ) -> Result<Page<Message>, DbError>;
-
-    /// Insert or update a single message
-    async fn insert(&self, message: &Message) -> Result<(), DbError>;
-
-    /// Insert or update multiple messages in a transaction
-    async fn insert_many(&self, messages: Vec<Message>) -> Result<(), DbError>;
-
-    /// Retrieve a single message by its ID
-    async fn get_by_id(&self, message_id: &str) -> Result<Option<Message>, DbError>;
-
-    /// Retrieve multiple messages by their IDs
-    async fn get_by_ids(&self, message_ids: Vec<String>) -> Result<Vec<Message>, DbError>;
-
-    /// Delete a single message by its ID
-    async fn delete_by_id(&self, message_id: &str) -> Result<(), DbError>;
-
-    /// Delete all messages for a specific session
-    async fn delete_by_session(&self, session_id: &str) -> Result<(), DbError>;
-
-    /// Update index metadata after rebuilding
-    async fn update_index_meta(
-        &self,
-        session_id: &str,
-        index_path: &str,
-        doc_count: usize,
-        rebuild_duration_ms: i64,
-    ) -> Result<(), DbError>;
-
-    /// Get the last indexed timestamp for a session
-    async fn get_last_indexed_at(&self, session_id: &str) -> Result<i64, DbError>;
-
-    /// Check if a session has messages newer than the last index build
-    async fn is_index_dirty(&self, session_id: &str) -> Result<bool, DbError>;
-
-    /// Delete index metadata for a specific session
-    async fn delete_index_metadata(&self, session_id: &str) -> Result<(), DbError>;
-
-    /// Get recent messages for a specific session with limit
-    async fn get_messages_by_session(
-        &self,
-        session_id: &str,
-        limit: u64,
-    ) -> Result<Vec<Message>, DbError>;
-
-    /// Get the most recent messages for a session in ascending chronological order.
-    /// Returns `limit + 1` internally so callers can infer whether older messages exist.
-    async fn get_recent_slice(
-        &self,
-        session_id: &str,
-        limit: u64,
-    ) -> Result<MessageSlicePage, DbError>;
-
-    /// Get messages older than a row-id cursor for a session in ascending causal order.
-    /// Returned cursors still include `created_at` for UI metadata, but row_id is the ordering
-    /// truth so message history is resilient to clock skew or cross-layer timestamp drift.
-    async fn get_messages_before(
-        &self,
-        session_id: &str,
-        before_row_id: i64,
-        limit: u64,
-    ) -> Result<MessageSlicePage, DbError>;
-
-    /// Get recent messages across all sessions with limit
-    async fn get_recent_messages(&self, limit: u64) -> Result<Vec<Message>, DbError>;
-
-    /// Get all distinct session IDs that have messages
-    async fn get_distinct_sessions(&self) -> Result<Vec<String>, DbError>;
-
-    /// Get message counts grouped by session ID
-    async fn count_by_session(&self) -> Result<Vec<(String, u64)>, DbError>;
-
-    /// Get message models (raw SeaORM models) for search indexing
-    async fn get_message_models_by_session(
-        &self,
-        session_id: &str,
-        limit: u64,
-    ) -> Result<Vec<message::Model>, DbError>;
-
-    /// Get recent message models across all sessions for search indexing
-    async fn get_recent_message_models(&self, limit: u64) -> Result<Vec<message::Model>, DbError>;
-}
-
-/// SQLite implementation of MessageRepository using SeaORM
-#[derive(Debug)]
-pub struct SqliteMessageRepository {
-    db: DatabaseConnection,
-}
-
-impl SqliteMessageRepository {
-    /// Create a new SQLite message repository with the given database connection
-    pub fn new(db: DatabaseConnection) -> Self {
-        Self { db }
-    }
-
-    fn row_to_message_model(row: &QueryResult) -> Result<message::Model, DbError> {
-        Ok(message::Model {
-            id: row.try_get("", "id").map_err(DbError::SeaOrmQueryFailed)?,
-            session_id: row
-                .try_get("", "session_id")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            role: row
-                .try_get("", "role")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            content: row
-                .try_get("", "content")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            tool_calls: row
-                .try_get("", "tool_calls")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            tool_call_id: row
-                .try_get("", "tool_call_id")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            is_streaming: row
-                .try_get("", "is_streaming")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            thinking: row
-                .try_get("", "thinking")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            thinking_signature: row
-                .try_get("", "thinking_signature")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            assistant_id: row
-                .try_get("", "assistant_id")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            attachments: row
-                .try_get("", "attachments")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            tool_use: row
-                .try_get("", "tool_use")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            created_at: row
-                .try_get("", "created_at")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            updated_at: row
-                .try_get("", "updated_at")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            source: row
-                .try_get("", "source")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            error: row
-                .try_get("", "error")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            usage: row
-                .try_get("", "usage")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            prompt_tokens: row
-                .try_get("", "prompt_tokens")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-        })
-    }
-
-    fn row_to_cursor(row: &QueryResult) -> Result<MessagePaginationCursor, DbError> {
-        Ok(MessagePaginationCursor {
-            created_at: row
-                .try_get("", "created_at")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-            row_id: row
-                .try_get("", "cursor_rowid")
-                .map_err(DbError::SeaOrmQueryFailed)?,
-        })
-    }
-
-    fn row_to_message_with_cursor(row: &QueryResult) -> Result<MessageRowWithCursor, DbError> {
-        Ok(MessageRowWithCursor {
-            model: Self::row_to_message_model(row)?,
-            cursor: Self::row_to_cursor(row)?,
-        })
-    }
-
-    async fn query_slice_rows(
-        &self,
-        sql: &str,
-        values: Vec<sea_orm::Value>,
-    ) -> Result<Vec<MessageRowWithCursor>, DbError> {
-        let rows = self
-            .db
-            .query_all(Statement::from_sql_and_values(
-                DatabaseBackend::Sqlite,
-                sql,
-                values,
-            ))
-            .await
-            .map_err(DbError::SeaOrmQueryFailed)?;
-
-        rows.iter()
-            .map(Self::row_to_message_with_cursor)
-            .collect::<Result<Vec<_>, _>>()
-    }
-
-    async fn query_message_models(
-        &self,
-        sql: &str,
-        values: Vec<sea_orm::Value>,
-    ) -> Result<Vec<message::Model>, DbError> {
-        let rows = self
-            .db
-            .query_all(Statement::from_sql_and_values(
-                DatabaseBackend::Sqlite,
-                sql,
-                values,
-            ))
-            .await
-            .map_err(DbError::SeaOrmQueryFailed)?;
-
-        rows.iter()
-            .map(Self::row_to_message_model)
-            .collect::<Result<Vec<_>, _>>()
-    }
-
-    fn validate_slice_limit(limit: u64) -> Result<i64, DbError> {
-        if limit == 0 {
-            return Err(DbError::InvalidInput(
-                "Message slice limit must be greater than zero".to_string(),
-            ));
-        }
-
-        i64::try_from(limit.saturating_add(1))
-            .map_err(|_| DbError::InvalidInput("Message slice limit is too large".to_string()))
-    }
-
-    fn build_slice_page(
-        mut rows: Vec<MessageRowWithCursor>,
-        limit: u64,
-    ) -> Result<MessageSlicePage, DbError> {
-        let has_more_before = rows.len() as u64 > limit;
-        if has_more_before {
-            rows.truncate(limit as usize);
-        }
-        rows.reverse();
-
-        let oldest_cursor = rows.first().map(|row| row.cursor.clone());
-        let items = rows
-            .into_iter()
-            .map(|row| Self::model_to_message(row.model))
-            .collect();
-
-        Ok(MessageSlicePage {
-            items,
-            has_more_before,
-            oldest_cursor,
-        })
-    }
-
-    /// Rebuild UI-facing metadata / error fields from persisted columns.
-    ///
-    /// `messages.metadata` is not a DB column. Tool UI state is persisted in the
-    /// `error` column as a JSON envelope:
-    /// - `{"toolError": true}` — tool failure marker (also legacy `content[].isError`)
-    /// - `{"structuredContent": ...}` — MCP structured_content for chat UI cards
-    /// - both keys may appear together
-    ///
-    /// When the column only holds this envelope, `Message.error` stays `None`
-    /// (that field is reserved for LLM/service failure UI payloads).
-    fn decode_persisted_tool_error(
-        content_json: &str,
-        error: Option<serde_json::Value>,
-    ) -> (Option<serde_json::Value>, Option<serde_json::Value>) {
-        let tool_error_from_legacy_content =
-            crate::mcp::types::raw_content_json_has_legacy_item_error(content_json);
-
-        let envelope = error.as_ref().and_then(|value| value.as_object());
-        let tool_error_from_column = envelope
-            .and_then(|object| object.get("toolError"))
-            .and_then(|value| value.as_bool())
-            .unwrap_or(false);
-        let structured_content = envelope
-            .and_then(|object| object.get("structuredContent"))
-            .cloned();
-        let is_ui_envelope = tool_error_from_column || structured_content.is_some();
-
-        let tool_error = tool_error_from_column || tool_error_from_legacy_content;
-
-        let metadata = {
-            let mut map = serde_json::Map::new();
-            if tool_error {
-                map.insert("toolError".to_string(), serde_json::Value::Bool(true));
-            }
-            if let Some(structured) = structured_content {
-                map.insert("structuredContent".to_string(), structured);
-            }
-            if map.is_empty() {
-                None
-            } else {
-                Some(serde_json::Value::Object(map))
-            }
-        };
-
-        // UI envelope occupies the error column; do not expose it as Message.error.
-        let error = if is_ui_envelope { None } else { error };
-
-        (error, metadata)
-    }
-
-    /// Encode in-memory tool UI metadata into the durable `error` column envelope.
-    ///
-    /// Persists `metadata.toolError` and/or `metadata.structuredContent` so chat
-    /// structured cards survive session reload. Falls back to `message.error` for
-    /// real LLM/service failure payloads when no UI envelope keys are present.
-    fn encode_persisted_error(message: &Message) -> Option<serde_json::Value> {
-        let tool_error = message
-            .metadata
-            .as_ref()
-            .and_then(|metadata| metadata.get("toolError"))
-            .and_then(|value| value.as_bool())
-            .unwrap_or(false);
-        let structured_content = message
-            .metadata
-            .as_ref()
-            .and_then(|metadata| metadata.get("structuredContent"))
-            .cloned();
-
-        if tool_error || structured_content.is_some() {
-            let mut map = serde_json::Map::new();
-            if tool_error {
-                map.insert("toolError".to_string(), serde_json::Value::Bool(true));
-            }
-            if let Some(structured) = structured_content {
-                map.insert("structuredContent".to_string(), structured);
-            }
-            return Some(serde_json::Value::Object(map));
-        }
-
-        message.error.clone()
-    }
-
-    /// Convert SeaORM message model to Message type
-    fn model_to_message(model: message::Model) -> Message {
-        let content: Vec<crate::mcp::types::MCPContent> = from_json_or_default(&model.content);
-
-        let tool_calls: Option<Vec<crate::agent::types::ToolCall>> =
-            from_json_option(&model.tool_calls);
-
-        let attachments: Option<serde_json::Value> = from_json_option(&model.attachments);
-
-        let tool_use: Option<serde_json::Value> = from_json_option(&model.tool_use);
-
-        let raw_error: Option<serde_json::Value> = from_json_option(&model.error);
-
-        let usage: Option<serde_json::Value> = from_json_option(&model.usage);
-
-        let (error, metadata) = Self::decode_persisted_tool_error(&model.content, raw_error);
-
-        Message {
-            id: model.id,
-            session_id: model.session_id,
-            role: model.role,
-            content,
-            tool_calls,
-            tool_call_id: model.tool_call_id,
-            is_streaming: model.is_streaming.map(|v| v != 0),
-            thinking: model.thinking,
-            thinking_signature: model.thinking_signature,
-            assistant_id: model.assistant_id,
-            attachments,
-            tool_use,
-            created_at: model.created_at,
-            updated_at: model.updated_at,
-            source: model.source.map(MessageSource::from_raw),
-            error,
-            usage,
-            prompt_tokens: model.prompt_tokens,
-            metadata,
-        }
-    }
-
-    /// Convert Message type to SeaORM ActiveModel
-    pub(crate) fn message_to_active_model(
-        message: &Message,
-    ) -> Result<message::ActiveModel, DbError> {
-        // Serialize structured types to JSON strings for DB storage
-        let content_json = serde_json::to_string(&message.content).map_err(|e| {
-            DbError::SerializationError(format!("Failed to serialize content: {}", e))
-        })?;
-
-        let tool_calls_json = to_json_option(&message.tool_calls).map_err(|e| {
-            DbError::SerializationError(format!("Failed to serialize tool_calls: {}", e))
-        })?;
-
-        let attachments_json = to_json_option(&message.attachments).map_err(|e| {
-            DbError::SerializationError(format!("Failed to serialize attachments: {}", e))
-        })?;
-
-        let tool_use_json = to_json_option(&message.tool_use).map_err(|e| {
-            DbError::SerializationError(format!("Failed to serialize tool_use: {}", e))
-        })?;
-
-        let error_json = to_json_option(&Self::encode_persisted_error(message)).map_err(|e| {
-            DbError::SerializationError(format!("Failed to serialize error: {}", e))
-        })?;
-
-        let usage_json = to_json_option(&message.usage).map_err(|e| {
-            DbError::SerializationError(format!("Failed to serialize usage: {}", e))
-        })?;
-
-        Ok(message::ActiveModel {
-            id: Set(message.id.clone()),
-            session_id: Set(message.session_id.clone()),
-            role: Set(message.role.clone()),
-            content: Set(content_json),
-            tool_calls: Set(tool_calls_json),
-            tool_call_id: Set(message.tool_call_id.clone()),
-            is_streaming: Set(message.is_streaming.map(|b| if b { 1 } else { 0 })),
-            thinking: Set(message.thinking.clone()),
-            thinking_signature: Set(message.thinking_signature.clone()),
-            assistant_id: Set(message.assistant_id.clone()),
-            attachments: Set(attachments_json),
-            tool_use: Set(tool_use_json),
-            created_at: Set(message.created_at),
-            updated_at: Set(message.updated_at),
-            source: Set(message
-                .source
-                .as_ref()
-                .map(|source| source.as_str().to_string())),
-            error: Set(error_json),
-            usage: Set(usage_json),
-            prompt_tokens: Set(message.prompt_tokens),
-        })
-    }
-
-    /// Helper to get the OnConflict strategy for upserting messages
-    pub(crate) fn get_upsert_on_conflict() -> sea_orm::sea_query::OnConflict {
-        use sea_orm::sea_query::OnConflict;
-        OnConflict::column(message::Column::Id)
-            .update_columns([
-                message::Column::SessionId,
-                message::Column::Role,
-                message::Column::Content,
-                message::Column::ToolCalls,
-                message::Column::ToolCallId,
-                message::Column::IsStreaming,
-                message::Column::Thinking,
-                message::Column::ThinkingSignature,
-                message::Column::AssistantId,
-                message::Column::Attachments,
-                message::Column::ToolUse,
-                message::Column::UpdatedAt,
-                message::Column::Source,
-                message::Column::Error,
-                message::Column::Usage,
-                message::Column::PromptTokens,
-            ])
-            .to_owned()
-    }
-
-    pub(crate) async fn update_session_last_message_at<C>(
-        db: &C,
-        session_id: &str,
-        last_message_at: i64,
-    ) -> Result<(), DbError>
-    where
-        C: ConnectionTrait,
-    {
-        session::Entity::update_many()
-            .col_expr(
-                session::Column::LastMessageAt,
-                Expr::cust_with_values("MAX(COALESCE(last_message_at, 0), ?)", [last_message_at]),
-            )
-            .filter(session::Column::Id.eq(session_id))
-            .exec(db)
-            .await?;
-
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl MessageRepository for SqliteMessageRepository {
-    async fn get_page(
-        &self,
-        session_id: &str,
-        page: u64,
-        page_size: u64,
-    ) -> Result<Page<Message>, DbError> {
-        if page_size == 0 {
-            return Err(DbError::InvalidInput("page_size must be > 0".into()));
-        }
-
-        // Get total count using SeaORM
-        let total = MessageEntity::find()
-            .filter(message::Column::SessionId.eq(session_id))
-            .count(&self.db)
-            .await?;
-
-        // Calculate offset
-        let offset = page.saturating_sub(1).saturating_mul(page_size);
-
-        // Fetch paginated messages in persisted causal order. Using rowid avoids treating
-        // cross-layer created_at skew as conversation truth.
-        let models = self
-            .query_message_models(
-                "SELECT id, session_id, role, content, tool_calls, tool_call_id, is_streaming, thinking, thinking_signature, assistant_id, attachments, tool_use, created_at, updated_at, source, error, usage, prompt_tokens \
-                 FROM messages \
-                 WHERE session_id = ? \
-                 ORDER BY rowid ASC \
-                 LIMIT ? OFFSET ?",
-                vec![session_id.into(), (page_size as i64).into(), (offset as i64).into()],
-            )
-            .await?;
-
-        let messages: Vec<Message> = models.into_iter().map(Self::model_to_message).collect();
-
-        Ok(Page::new(messages, page, page_size, total))
-    }
-
-    async fn insert(&self, message: &Message) -> Result<(), DbError> {
-        let model = Self::message_to_active_model(message)?;
-
-        MessageEntity::insert(model)
-            .on_conflict(Self::get_upsert_on_conflict())
-            .exec(&self.db)
-            .await?;
-
-        Self::update_session_last_message_at(&self.db, &message.session_id, message.created_at)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn insert_many(&self, messages: Vec<Message>) -> Result<(), DbError> {
-        use sea_orm::TransactionTrait;
-        use std::collections::HashMap;
-
-        let txn = self.db.begin().await?;
-        let mut latest_by_session: HashMap<String, i64> = HashMap::new();
-
-        for message in messages {
-            let model = Self::message_to_active_model(&message)?;
-
-            MessageEntity::insert(model)
-                .on_conflict(Self::get_upsert_on_conflict())
-                .exec(&txn)
-                .await?;
-
-            latest_by_session
-                .entry(message.session_id.clone())
-                .and_modify(|current| *current = (*current).max(message.created_at))
-                .or_insert(message.created_at);
-        }
-
-        for (session_id, last_message_at) in latest_by_session {
-            Self::update_session_last_message_at(&txn, &session_id, last_message_at).await?;
-        }
-
-        txn.commit()
-            .await
-            .map_err(|e| DbError::TransactionFailed(e.to_string()))?;
-        Ok(())
-    }
-
-    async fn get_by_id(&self, message_id: &str) -> Result<Option<Message>, DbError> {
-        let model = MessageEntity::find_by_id(message_id).one(&self.db).await?;
-        Ok(model.map(Self::model_to_message))
-    }
-
-    async fn get_by_ids(&self, message_ids: Vec<String>) -> Result<Vec<Message>, DbError> {
-        if message_ids.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        let models = MessageEntity::find()
-            .filter(message::Column::Id.is_in(message_ids.clone()))
-            .all(&self.db)
-            .await?;
-
-        // Map results for efficient lookup
-        let mut msg_map: std::collections::HashMap<String, Message> = models
-            .into_iter()
-            .map(|m| (m.id.clone(), Self::model_to_message(m)))
-            .collect();
-
-        // Sort to match requested order
-        let result = message_ids
-            .into_iter()
-            .filter_map(|id| msg_map.remove(&id))
-            .collect();
-
-        Ok(result)
-    }
-
-    async fn delete_by_id(&self, message_id: &str) -> Result<(), DbError> {
-        MessageEntity::delete_by_id(message_id)
-            .exec(&self.db)
-            .await?;
-        Ok(())
-    }
-
-    async fn delete_by_session(&self, session_id: &str) -> Result<(), DbError> {
-        MessageEntity::delete_many()
-            .filter(message::Column::SessionId.eq(session_id))
-            .exec(&self.db)
-            .await?;
-        Ok(())
-    }
-
-    async fn update_index_meta(
-        &self,
-        session_id: &str,
-        index_path: &str,
-        doc_count: usize,
-        rebuild_duration_ms: i64,
-    ) -> Result<(), DbError> {
-        use sea_orm::sea_query::OnConflict;
-        let now = chrono::Utc::now().timestamp_millis();
-
-        let model = message_index_meta::ActiveModel {
-            session_id: Set(session_id.to_string()),
-            index_path: Set(Some(index_path.to_string())),
-            last_indexed_at: Set(now),
-            doc_count: Set(doc_count as i32),
-            index_version: Set(1),
-            last_rebuild_duration_ms: Set(Some(rebuild_duration_ms)),
-        };
-
-        MessageIndexMeta::insert(model)
-            .on_conflict(
-                OnConflict::column(message_index_meta::Column::SessionId)
-                    .update_columns([
-                        message_index_meta::Column::IndexPath,
-                        message_index_meta::Column::LastIndexedAt,
-                        message_index_meta::Column::DocCount,
-                        message_index_meta::Column::LastRebuildDurationMs,
-                    ])
-                    .to_owned(),
-            )
-            .exec(&self.db)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn get_last_indexed_at(&self, session_id: &str) -> Result<i64, DbError> {
-        let result = MessageIndexMeta::find_by_id(session_id)
-            .one(&self.db)
-            .await?;
-
-        Ok(result.map(|m| m.last_indexed_at).unwrap_or(0))
-    }
-
-    async fn is_index_dirty(&self, session_id: &str) -> Result<bool, DbError> {
-        use sea_orm::sea_query::Expr;
-
-        let last_indexed_at = self.get_last_indexed_at(session_id).await?;
-
-        let max_created: Option<i64> = MessageEntity::find()
-            .filter(message::Column::SessionId.eq(session_id))
-            .select_only()
-            .column_as(Expr::col(message::Column::CreatedAt).max(), "max_created")
-            .into_tuple()
-            .one(&self.db)
-            .await?;
-
-        Ok(max_created.map(|t| t > last_indexed_at).unwrap_or(false))
-    }
-
-    async fn delete_index_metadata(&self, session_id: &str) -> Result<(), DbError> {
-        MessageIndexMeta::delete_by_id(session_id)
-            .exec(&self.db)
-            .await?;
-        Ok(())
-    }
-
-    async fn get_messages_by_session(
-        &self,
-        session_id: &str,
-        limit: u64,
-    ) -> Result<Vec<Message>, DbError> {
-        let models = self
-            .get_message_models_by_session(session_id, limit)
-            .await?;
-        Ok(models.into_iter().map(Self::model_to_message).collect())
-    }
-
-    async fn get_recent_slice(
-        &self,
-        session_id: &str,
-        limit: u64,
-    ) -> Result<MessageSlicePage, DbError> {
-        let fetch_limit = Self::validate_slice_limit(limit)?;
-        let rows = self
-            .query_slice_rows(
-                "SELECT rowid AS cursor_rowid, id, session_id, role, content, tool_calls, tool_call_id, is_streaming, thinking, thinking_signature, assistant_id, attachments, tool_use, created_at, updated_at, source, error, usage, prompt_tokens \
-                 FROM messages \
-                 WHERE session_id = ? \
-                 ORDER BY rowid DESC \
-                 LIMIT ?",
-                vec![session_id.into(), fetch_limit.into()],
-            )
-            .await?;
-
-        Self::build_slice_page(rows, limit)
-    }
-
-    async fn get_messages_before(
-        &self,
-        session_id: &str,
-        before_row_id: i64,
-        limit: u64,
-    ) -> Result<MessageSlicePage, DbError> {
-        let fetch_limit = Self::validate_slice_limit(limit)?;
-        let rows = self
-            .query_slice_rows(
-                "SELECT rowid AS cursor_rowid, id, session_id, role, content, tool_calls, tool_call_id, is_streaming, thinking, thinking_signature, assistant_id, attachments, tool_use, created_at, updated_at, source, error, usage, prompt_tokens \
-                 FROM messages \
-                 WHERE session_id = ? \
-                   AND rowid < ? \
-                 ORDER BY rowid DESC \
-                 LIMIT ?",
-                vec![
-                    session_id.into(),
-                    before_row_id.into(),
-                    fetch_limit.into(),
-                ],
-            )
-            .await?;
-
-        Self::build_slice_page(rows, limit)
-    }
-
-    async fn get_recent_messages(&self, limit: u64) -> Result<Vec<Message>, DbError> {
-        let models = self.get_recent_message_models(limit).await?;
-        Ok(models.into_iter().map(Self::model_to_message).collect())
-    }
-
-    async fn get_distinct_sessions(&self) -> Result<Vec<String>, DbError> {
-        let sessions: Vec<String> = MessageEntity::find()
-            .select_only()
-            .column(message::Column::SessionId)
-            .distinct()
-            .into_tuple()
-            .all(&self.db)
-            .await?;
-
-        Ok(sessions)
-    }
-
-    async fn count_by_session(&self) -> Result<Vec<(String, u64)>, DbError> {
-        let rows: Vec<(String, i64)> = MessageEntity::find()
-            .select_only()
-            .column(message::Column::SessionId)
-            .column_as(Expr::col(message::Column::Id).count(), "message_count")
-            .group_by(message::Column::SessionId)
-            .into_tuple()
-            .all(&self.db)
-            .await?;
-
-        Ok(rows
-            .into_iter()
-            .map(|(session_id, count)| (session_id, count.max(0) as u64))
-            .collect())
-    }
-
-    async fn get_message_models_by_session(
-        &self,
-        session_id: &str,
-        limit: u64,
-    ) -> Result<Vec<message::Model>, DbError> {
-        let models = self
-            .query_message_models(
-                "SELECT id, session_id, role, content, tool_calls, tool_call_id, is_streaming, thinking, thinking_signature, assistant_id, attachments, tool_use, created_at, updated_at, source, error, usage, prompt_tokens \
-                 FROM messages \
-                 WHERE session_id = ? \
-                 ORDER BY rowid DESC \
-                 LIMIT ?",
-                vec![session_id.into(), (limit as i64).into()],
-            )
-            .await?;
-
-        Ok(models)
-    }
-
-    async fn get_recent_message_models(&self, limit: u64) -> Result<Vec<message::Model>, DbError> {
-        let models = self
-            .query_message_models(
-                "SELECT id, session_id, role, content, tool_calls, tool_call_id, is_streaming, thinking, thinking_signature, assistant_id, attachments, tool_use, created_at, updated_at, source, error, usage, prompt_tokens \
-                 FROM messages \
-                 ORDER BY rowid DESC \
-                 LIMIT ?",
-                vec![(limit as i64).into()],
-            )
-            .await?;
-
-        Ok(models)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::entity::prelude::Session as SessionEntity;
-    use crate::entity::session;
-    use crate::mcp::types::MCPContent;
-
-    async fn setup_test_db() -> SqliteMessageRepository {
-        crate::lifecycle::database::register_sqlite_vec();
-        let db = sea_orm::Database::connect("sqlite::memory:")
-            .await
-            .expect("Failed to create in-memory database");
-
-        // Run migrations
-        use migration::{Migrator, MigratorTrait};
-        Migrator::up(&db, None)
-            .await
-            .expect("Failed to run migrations");
-
-        SqliteMessageRepository::new(db)
-    }
-
-    async fn create_test_session(db: &DatabaseConnection, session_id: &str) {
-        let now = chrono::Utc::now().timestamp_millis();
-        let session = session::ActiveModel {
-            id: Set(session_id.to_string()),
-            name: Set(Some("Test Session".to_string())),
-            status: Set("idle".to_string()),
-            model: Set("gpt-4".to_string()),
-            provider: Set("openai".to_string()),
-            execution_mode: Set("normal".to_string()),
-            is_bookmarked: Set(false),
-            created_at: Set(now),
-            updated_at: Set(now),
-            ..Default::default()
-        };
-        SessionEntity::insert(session)
-            .exec(db)
-            .await
-            .expect("Failed to create session");
-    }
-
-    fn create_dummy_message(id: &str, session_id: &str) -> Message {
-        Message {
-            id: id.to_string(),
-            session_id: session_id.to_string(),
-            role: "user".to_string(),
-            content: vec![MCPContent::Text {
-                text: "Hello".to_string(),
-            }],
-            tool_calls: None,
-            tool_call_id: None,
-            is_streaming: Some(false),
-            thinking: None,
-            thinking_signature: None,
-            assistant_id: None,
-            attachments: None,
-            tool_use: None,
-            created_at: 1000,
-            updated_at: 1000,
-            source: None,
-            error: None,
-            metadata: None,
-            usage: None,
-            prompt_tokens: None,
-        }
-    }
-
-    #[test]
-    fn encode_decode_preserves_structured_content_envelope() {
-        let structured = serde_json::json!({
-            "sessionId": "a1b2c3d4e5",
-            "status": "started",
-            "responseStatus": "pending",
-        });
-        let mut message = create_dummy_message("tool-structured", "session1");
-        message.role = "tool".to_string();
-        message.metadata = Some(serde_json::json!({
-            "structuredContent": structured,
-        }));
-
-        let encoded = SqliteMessageRepository::encode_persisted_error(&message);
-        assert_eq!(
-            encoded
-                .as_ref()
-                .and_then(|value| value.get("structuredContent")),
-            Some(&structured)
-        );
-
-        let (error, metadata) = SqliteMessageRepository::decode_persisted_tool_error("[]", encoded);
-        assert!(error.is_none());
-        assert_eq!(
-            metadata
-                .as_ref()
-                .and_then(|value| value.get("structuredContent")),
-            Some(&structured)
-        );
-    }
-
-    #[test]
-    fn encode_decode_preserves_tool_error_and_structured_content() {
-        let structured = serde_json::json!({ "path": "a.txt", "action": "created" });
-        let mut message = create_dummy_message("tool-both", "session1");
-        message.metadata = Some(serde_json::json!({
-            "toolError": true,
-            "structuredContent": structured,
-        }));
-
-        let encoded = SqliteMessageRepository::encode_persisted_error(&message);
-        let (error, metadata) = SqliteMessageRepository::decode_persisted_tool_error("[]", encoded);
-
-        assert!(error.is_none());
-        assert_eq!(
-            metadata
-                .as_ref()
-                .and_then(|value| value.get("toolError"))
-                .and_then(|value| value.as_bool()),
-            Some(true)
-        );
-        assert_eq!(
-            metadata
-                .as_ref()
-                .and_then(|value| value.get("structuredContent")),
-            Some(&structured)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_insert_and_get_messages() {
-        let repo = setup_test_db().await;
-        create_test_session(&repo.db, "session1").await;
-        let message = create_dummy_message("msg1", "session1");
-
-        repo.insert(&message).await.expect("Failed to insert");
-
-        let messages = repo
-            .get_messages_by_session("session1", 10)
-            .await
-            .expect("Failed to get messages");
-
-        assert_eq!(messages.len(), 1);
-        assert_eq!(messages[0].id, "msg1");
-    }
-
-    #[tokio::test]
-    async fn test_reload_derives_tool_error_from_error_column_marker() {
-        let repo = setup_test_db().await;
-        create_test_session(&repo.db, "session1").await;
-
-        let mut message = create_dummy_message("tool-err", "session1");
-        message.role = "tool".to_string();
-        message.tool_call_id = Some("call-1".to_string());
-        message.content = vec![MCPContent::Text {
-            text: "tool failed".to_string(),
-        }];
-        message.metadata = Some(serde_json::json!({ "toolError": true }));
-
-        repo.insert(&message).await.expect("Failed to insert");
-
-        let loaded = repo
-            .get_by_id("tool-err")
-            .await
-            .expect("Failed to get message")
-            .expect("message should exist");
-
-        assert_eq!(
-            loaded
-                .metadata
-                .as_ref()
-                .and_then(|metadata| metadata.get("toolError"))
-                .and_then(|value| value.as_bool()),
-            Some(true),
-            "toolError must be reconstructed from persisted error-column marker"
-        );
-        assert!(
-            loaded.error.is_none(),
-            "toolError marker must not surface as Message.error"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_reload_preserves_structured_content_in_error_column_envelope() {
-        let repo = setup_test_db().await;
-        create_test_session(&repo.db, "session1").await;
-
-        let structured = serde_json::json!({
-            "sessionId": "a1b2c3d4e5",
-            "status": "started",
-            "responseStatus": "pending",
-            "toolName": "startSession",
-        });
-
-        let mut message = create_dummy_message("tool-structured", "session1");
-        message.role = "tool".to_string();
-        message.tool_call_id = Some("call-spawn".to_string());
-        message.content = vec![MCPContent::Text {
-            text: "Session started successfully".to_string(),
-        }];
-        message.metadata = Some(serde_json::json!({
-            "structuredContent": structured,
-        }));
-
-        repo.insert(&message).await.expect("Failed to insert");
-
-        let loaded = repo
-            .get_by_id("tool-structured")
-            .await
-            .expect("Failed to get message")
-            .expect("message should exist");
-
-        assert_eq!(
-            loaded
-                .metadata
-                .as_ref()
-                .and_then(|metadata| metadata.get("structuredContent")),
-            Some(&structured),
-            "structuredContent must survive DB reload for structured tool cards"
-        );
-        assert!(
-            loaded.error.is_none(),
-            "structuredContent envelope must not surface as Message.error"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_reload_preserves_structured_content_with_tool_error() {
-        let repo = setup_test_db().await;
-        create_test_session(&repo.db, "session1").await;
-
-        let structured = serde_json::json!({
-            "path": "a.txt",
-            "action": "created",
-        });
-
-        let mut message = create_dummy_message("tool-both", "session1");
-        message.role = "tool".to_string();
-        message.tool_call_id = Some("call-both".to_string());
-        message.content = vec![MCPContent::Text {
-            text: "write failed".to_string(),
-        }];
-        message.metadata = Some(serde_json::json!({
-            "toolError": true,
-            "structuredContent": structured,
-        }));
-
-        repo.insert(&message).await.expect("Failed to insert");
-
-        let loaded = repo
-            .get_by_id("tool-both")
-            .await
-            .expect("Failed to get message")
-            .expect("message should exist");
-
-        assert_eq!(
-            loaded
-                .metadata
-                .as_ref()
-                .and_then(|metadata| metadata.get("toolError"))
-                .and_then(|value| value.as_bool()),
-            Some(true)
-        );
-        assert_eq!(
-            loaded
-                .metadata
-                .as_ref()
-                .and_then(|metadata| metadata.get("structuredContent")),
-            Some(&structured)
-        );
-        assert!(loaded.error.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_reload_derives_tool_error_from_legacy_content_json_is_error() {
-        let repo = setup_test_db().await;
-        create_test_session(&repo.db, "session1").await;
-
-        // Simulate a pre-cleanup DB row that still has content[].isError.
-        use sea_orm::{ActiveModelTrait, Set};
-        let now = chrono::Utc::now().timestamp_millis();
-        crate::entity::message::ActiveModel {
-            id: Set("legacy-tool-err".to_string()),
-            session_id: Set("session1".to_string()),
-            role: Set("tool".to_string()),
-            content: Set(r#"[{"type":"text","text":"legacy fail","isError":true}]"#.to_string()),
-            tool_calls: Set(None),
-            tool_call_id: Set(Some("call-legacy".to_string())),
-            is_streaming: Set(Some(0)),
-            thinking: Set(None),
-            thinking_signature: Set(None),
-            assistant_id: Set(None),
-            attachments: Set(None),
-            tool_use: Set(None),
-            created_at: Set(now),
-            updated_at: Set(now),
-            source: Set(Some("tool".to_string())),
-            error: Set(None),
-            usage: Set(None),
-            prompt_tokens: Set(None),
-        }
-        .insert(&repo.db)
-        .await
-        .expect("Failed to insert legacy row");
-
-        let loaded = repo
-            .get_by_id("legacy-tool-err")
-            .await
-            .expect("Failed to get message")
-            .expect("message should exist");
-
-        assert_eq!(
-            loaded
-                .metadata
-                .as_ref()
-                .and_then(|metadata| metadata.get("toolError"))
-                .and_then(|value| value.as_bool()),
-            Some(true),
-            "toolError must be reconstructed from legacy content JSON isError"
-        );
-        // Typed content must not retain the removed field (serde ignores unknown keys).
-        assert!(matches!(
-            loaded.content.first(),
-            Some(MCPContent::Text { text }) if text == "legacy fail"
-        ));
-    }
-
-    #[tokio::test]
-    async fn test_insert_many() {
-        let repo = setup_test_db().await;
-        create_test_session(&repo.db, "session1").await;
-        let messages = vec![
-            create_dummy_message("msg1", "session1"),
-            create_dummy_message("msg2", "session1"),
-        ];
-
-        repo.insert_many(messages)
-            .await
-            .expect("Failed to insert many");
-
-        let messages = repo
-            .get_messages_by_session("session1", 10)
-            .await
-            .expect("Failed to get messages");
-
-        assert_eq!(messages.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn test_get_recent_messages() {
-        let repo = setup_test_db().await;
-        create_test_session(&repo.db, "session1").await;
-        create_test_session(&repo.db, "session2").await;
-
-        let mut msg1 = create_dummy_message("msg1", "session1");
-        msg1.created_at = 1000;
-        let mut msg2 = create_dummy_message("msg2", "session2");
-        msg2.created_at = 2000;
-
-        repo.insert(&msg1).await.expect("Failed to insert");
-        repo.insert(&msg2).await.expect("Failed to insert");
-
-        let recent = repo
-            .get_recent_messages(10)
-            .await
-            .expect("Failed to get recent");
-        assert_eq!(recent.len(), 2);
-        assert_eq!(recent[0].id, "msg2"); // msg2 is newer
-    }
-
-    #[tokio::test]
-    async fn test_get_page_uses_persisted_row_order_when_created_at_is_inverted() {
-        let repo = setup_test_db().await;
-        create_test_session(&repo.db, "session1").await;
-
-        let mut assistant = create_dummy_message("assistant-owner", "session1");
-        assistant.created_at = 2000;
-        assistant.updated_at = 2000;
-        repo.insert(&assistant)
-            .await
-            .expect("Failed to insert assistant");
-
-        let mut tool_a = create_dummy_message("tool-result-a", "session1");
-        tool_a.created_at = 1000;
-        tool_a.updated_at = 1000;
-        repo.insert(&tool_a).await.expect("Failed to insert tool A");
-
-        let mut tool_b = create_dummy_message("tool-result-b", "session1");
-        tool_b.created_at = 1001;
-        tool_b.updated_at = 1001;
-        repo.insert(&tool_b).await.expect("Failed to insert tool B");
-
-        let page = repo
-            .get_page("session1", 1, 10)
-            .await
-            .expect("Failed to get page");
-        let ids: Vec<String> = page.items.into_iter().map(|message| message.id).collect();
-        assert_eq!(
-            ids,
-            vec![
-                "assistant-owner".to_string(),
-                "tool-result-a".to_string(),
-                "tool-result-b".to_string()
-            ]
-        );
-    }
-}
